+++ stopDatabox.sh
@@ -4,3 +4,7 @@ docker stack rm databox
 docker service rm $(docker service ls -q)
 
 docker secret rm $(docker secret ls -q)
+
+echo "waiting ....."
+sleep 10 #give docker some time to remove the networks etc
+echo "done."